+++ src/compiler/output-targets/dist-collection/index.ts
@@ -1,5 +1,5 @@
 import type * as d from '../../../declarations';
-import { catchError } from '@utils';
+import { catchError, generatePreamble } from '@utils';
 import { isOutputTargetDistCollection } from '../output-utils';
 import { join, relative } from 'path';
 import { writeCollectionManifests } from '../output-collection';
@@ -20,7 +20,10 @@ export const outputCollection = async (
   try {
     await Promise.all(
       changedModuleFiles.map(async (mod) => {
-        const code = mod.staticSourceFileText;
+        let code = mod.staticSourceFileText;
+        if (config.preamble) {
+          code = `${generatePreamble(config)}\n${code}`;
+        }
         const mapCode = mod.sourceMapFileText;
 
         await Promise.all(