+++ src/compiler/output-targets/dist-custom-elements/index.ts
@@ -4,16 +4,16 @@ import { bundleOutput } from '../../bundle/bundle-output';
 import {
   catchError,
   dashToPascalCase,
-  formatComponentRuntimeMeta,
+  generatePreamble,
   getSourceMappingUrlForEndOfFile,
   hasError,
   rollupToStencilSourceMap,
-  stringifyRuntimeData,
 } from '@utils';
 import { getCustomElementsBuildConditionals } from '../dist-custom-elements-bundle/custom-elements-build-conditionals';
 import { isOutputTargetDistCustomElements } from '../output-utils';
 import { join } from 'path';
 import { nativeComponentTransform } from '../../transformers/component-native/tranform-to-native-component';
+import { addDefineCustomElementFunctions } from '../../transformers/component-native/add-define-custom-element-function';
 import { optimizeModule } from '../../optimize/optimize-module';
 import { removeCollectionImports } from '../../transformers/remove-collection-imports';
 import { STENCIL_INTERNAL_CLIENT_ID, USER_INDEX_ENTRY_ID, STENCIL_APP_GLOBALS_ID } from '../../bundle/entry-alias-ids';
@@ -52,7 +52,12 @@ const bundleCustomElements = async (
       id: 'customElements',
       platform: 'client',
       conditionals: getCustomElementsBuildConditionals(config, buildCtx.components),
-      customTransformers: getCustomElementBundleCustomTransformer(config, compilerCtx),
+      customTransformers: getCustomElementBundleCustomTransformer(
+        config,
+        compilerCtx,
+        buildCtx.components,
+        outputTarget
+      ),
       externalRuntime: !!outputTarget.externalRuntime,
       inlineWorkers: true,
       inputs: {
@@ -70,6 +75,7 @@ const bundleCustomElements = async (
     const build = await bundleOutput(config, compilerCtx, buildCtx, bundleOpts);
     if (build) {
       const rollupOutput = await build.generate({
+        banner: generatePreamble(config),
         format: 'esm',
         sourcemap: config.sourceMap,
         chunkFileNames: outputTarget.externalRuntime || !config.hashFileNames ? '[name].js' : 'p-[hash].js',
@@ -83,6 +89,7 @@ const bundleCustomElements = async (
         if (bundle.type === 'chunk') {
           let code = bundle.code;
           let sourceMap = rollupToStencilSourceMap(bundle.map);
+
           const optimizeResults = await optimizeModule(config, compilerCtx, {
             input: code,
             isCore: bundle.isEntry,
@@ -132,11 +139,11 @@ const addCustomElementInputs = (
     if (cmp.isPlain) {
       exp.push(`export { ${importName} as ${exportName} } from '${cmp.sourceFilePath}';`);
     } else {
-      const meta = stringifyRuntimeData(formatComponentRuntimeMeta(cmp, false));
-
-      exp.push(`import { proxyCustomElement } from '${STENCIL_INTERNAL_CLIENT_ID}';`);
-      exp.push(`import { ${importName} as ${importAs} } from '${cmp.sourceFilePath}';`);
-      exp.push(`export const ${exportName} = /*@__PURE__*/proxyCustomElement(${importAs}, ${meta});`);
+      exp.push(
+        `import { ${importName} as ${importAs}, defineCustomElement as cmpDefCustomEle } from '${cmp.sourceFilePath}';`
+      );
+      exp.push(`export const ${exportName} = ${importAs};`);
+      exp.push(`export const defineCustomElement = cmpDefCustomEle;`);
     }
 
     bundleOpts.inputs[cmp.tagName] = coreKey;
@@ -160,7 +167,12 @@ const generateEntryPoint = (outputTarget: d.OutputTargetDistCustomElements, _bui
   return [...imp, ...exp].join('\n') + '\n';
 };
 
-const getCustomElementBundleCustomTransformer = (config: d.Config, compilerCtx: d.CompilerCtx) => {
+const getCustomElementBundleCustomTransformer = (
+  config: d.Config,
+  compilerCtx: d.CompilerCtx,
+  components: d.ComponentCompilerMeta[],
+  outputTarget: d.OutputTargetDistCustomElements
+) => {
   const transformOpts: d.TransformOptions = {
     coreImportPath: STENCIL_INTERNAL_CLIENT_ID,
     componentExport: null,
@@ -171,6 +183,7 @@ const getCustomElementBundleCustomTransformer = (config: d.Config, compilerCtx: 
     styleImportData: 'queryparams',
   };
   return [
+    addDefineCustomElementFunctions(compilerCtx, components, outputTarget),
     updateStencilCoreImports(transformOpts.coreImportPath),
     nativeComponentTransform(compilerCtx, transformOpts),
     removeCollectionImports(compilerCtx),