+++ src/utils/index.ts
@@ -12,4 +12,5 @@ export * from './normalize-path';
 export * from './url-paths';
 export * from './util';
 export * from './validation';
+export * from './byte-size';
 export * from './sourcemaps';